Move DisplaySettings to its package.
