+++ src/main/java/fiji/plugin/trackmate/features/TrackFeatureGrapher.java
@@ -21,7 +21,7 @@ import org.jfree.data.xy.XYSeriesCollection;
 import fiji.plugin.trackmate.Dimension;
 import fiji.plugin.trackmate.FeatureModel;
 import fiji.plugin.trackmate.Model;
-import fiji.plugin.trackmate.gui.DisplaySettings;
+import fiji.plugin.trackmate.gui.displaysettings.DisplaySettings;
 import fiji.plugin.trackmate.util.ExportableChartPanel;
 import fiji.plugin.trackmate.util.TMUtils;
