+++ src/main/java/fiji/plugin/trackmate/gui/descriptors/InitFilterDescriptor.java
@@ -9,8 +9,8 @@ import fiji.plugin.trackmate.Spot;
 import fiji.plugin.trackmate.TrackMate;
 import fiji.plugin.trackmate.features.FeatureFilter;
 import fiji.plugin.trackmate.features.FeatureUtils;
-import fiji.plugin.trackmate.gui.DisplaySettings.ObjectType;
 import fiji.plugin.trackmate.gui.TrackMateGUIController;
+import fiji.plugin.trackmate.gui.displaysettings.DisplaySettings.ObjectType;
 import fiji.plugin.trackmate.gui.panels.InitFilterPanel;
 
 public class InitFilterDescriptor implements WizardPanelDescriptor
