+++ src/test/java/fiji/plugin/trackmate/interactivetests/TrackLayout_Test.java
@@ -2,7 +2,7 @@ package fiji.plugin.trackmate.interactivetests;
 
 import fiji.plugin.trackmate.Model;
 import fiji.plugin.trackmate.SelectionModel;
-import fiji.plugin.trackmate.gui.DisplaySettings;
+import fiji.plugin.trackmate.gui.displaysettings.DisplaySettings;
 import fiji.plugin.trackmate.visualization.trackscheme.TrackScheme;
 
 public class TrackLayout_Test {
